Added connectivity functionality and button management
MCBrowser functionality so that the device can connect to other devices
- Still need to add functionality for the user to change the device ID
and the session ID

Added functionality for the manager and instructor buttons so the
selection can be saved out to the UserSingleton. Also changed the look
of the buttons so they would have rounded edges and change color when
selected.