Made the table view functional
Added the table view, and connected the delegates

Provided an array to hold the template items so they can be passed
around to other parts of the program.

Trying an alert controller to provide functionality for the add
function. Might not work out to benefit the project.

Added test items for the dataArray and the table view so that we have
something to test functional code with.

Added functions to handle the alert text fields so the values can be
more easily extracted.

Trying to utilize a mutable item for the add function.

Added a function which will clear out the text on the alert text fields.

Added a toggle switch of the template table view for deleting items.
Haven’t added the ability to remove them from the array, though.

Added functionality to the table view so it will load a custom
repeating cell that can be modified in code.