Changed functionality of the template VC and options VC
Template VC:

Moved the location of the mutableItem in the template view controller
to allow the program to update the array objects as they are added.
With the mutableItem declared outside the function as it was, the
updates to the template were adding the same instance of the object to
the array, then changing the information shown on all of the objects
instead of adding them individually.

Did proper research on how to update the Table View data, and decided
to utilize the reloadData functionality that is already built into the
protocol.

With the reloadData function, we don’t need the if (editingStyle)
portion that handles insertion of data, so it was removed.

Also added a toggle function that will toggle the editing value of the
table view AFTER an item has been deleted, so the user doesn’t have to
press the delete button every time they want to stop being able to
delete something.

Options VC:

Added the UITextFieldDelegate protocols to the main VC

Added a variable to hold the device name

Added an alert constant message to handle an error caused by leaving
the Session ID blank

Set the delegate of the text fields to the main view controller so the
data can be handled in code

Moved the advertiser for sessions into a function, instead of having
the VC start advertising on load, so the user can change the Session ID
if necessary.

Added functionality to change the values of the device name and
serviceID to match what is in the Text Fields when the return key is
pressed. Submitted a change request to the client to add a save button
so saving the information is more apparent.

Added error handling within the userConnect function.