Made numerous changes to the calendar view controller
Added labels, collection view outlets, segmented controls, a button to
share information, and a button to add appointments.

Added variables to make some custom reactive UI functionality to the
Collection View Cells.

Added functionality to hold the index path of a selected Collection
View Cell.

Set reactive restrictions to the calendar cells.

Added the necessary functions to make the collection view load and
work, with the ability to change the number and size of cells onscreen.

Added the ability to provide user feedback for a selected cell. - Still
needs error handling to prevent the user from selecting multiple cells
at one time.

Defined a custom function that will set the size of the cells based on
two floats.

Added a case switch to the Segmented control to handle switching
between day, week, and month views.

Added a condition to the share button that will check the user
singleton to see if the user has connected to a peer.