Custom Calendar Cell file
Added functionality to the calendar cell file, importing the UIKit to
allow creation of a label. This function will likely be moved in the
near future.

Currently working on solutions to add reactive layout design to the
label inside the collection view cell. This is so the user can add
appointment information to the cells. This will probably be removed and
placed in the main View Controller file and accessed inside one of the
Collection View Delegate functions.